Fix anti-snowball behavior on edge condition
I still don't understand conditions that can lead to this issue but here
is the behavior we intend to fix:
- single instance application
- upgradeStrategy is set to 0 for both maximumOverCapacity & minimumHealthCapacity
=> when task get unhealthy, it happens (from time to time) that:
- task is not killed
- we log in debug: "checkEnoughInstancesRunning: -1.0 >= 0.0"

The error came because we assumed (before this patch) that current
instance was still considered healthy thus: (healthyInstances - 1) is
the number of healthy instances remaining if we kill the current one.

"checkEnoughInstancesRunning: -1.0 >= 0.0" log line proves that it is
not always the case.

@komuta and I were not able to find any out of order execution that
would lead to this scenario but the current patch is more sane:
we now remove current instance from "healthy instances" list instead of
blindly removing 1 from the count of healthy instances.

Sadly no test can be written since I don't understand how it can be
triggered.

Change-Id: I182a23a9c708f81512682ba40a175957c2cd1d95