Always kill oldest instances first
Right now the order of task killing is "unhealthy" first.
For very big applications, killing unhealthy and then oldest tasks first
is a good heuristic. It favors killing first the instances that may have
have leaked.

Of course, we could find arguments for sorting in reverse (oldest
instacne have better caches) but at least we are specifying the order
instead of letting it random.

Change-Id: Ia4e7ec7751585f337d740a1246cac379d4871995
JIRA: MESOS-4343